// src/generated/model/programResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ProgramResponseStatus } from './programResponseStatus.ts';

export interface ProgramResponse {
  id?: string;
  code?: string;
  name?: string;
  description?: string;
  status?: ProgramResponseStatus;
  displayOrder?: number;
  createdAt?: string;
  updatedAt?: string;
}